requirements-dev.txt
# Development dependencies
-r requirements.txt

# Code quality and formatting
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
bandit==1.7.5
pre-commit==3.6.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.2.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# API documentation
drf-spectacular==0.27.0 